move tests to test-folder
