+++ .github/workflows/ci.yml
@@ -20,7 +20,7 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install -r requirements.txt
-      - name: Run tests with pytest
-        run: pytest 
-      - name: Run tests with pylint
+      - name: Run python tests with pytest
+        run: python -m pytest -vv tests/
+      - name: Run python linting with pylint
         run: find . -type f -name "*.py" | xargs pylint
