+++ tools/string_tools.py
